requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
dnspython==2.4.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.4.1
python-dotenv==1.0.0
Werkzeug==2.3.6
